relationships merged and regularized
