+++ models/hospital.py
@@ -36,13 +36,15 @@ class Hospital(BaseModel, Base):
     """
     __tablename__ = "hospital"
     hospitalID = Column(String, ForeignKey(
-        "admin.hospitalID",  ondelete="CASCADE"), nullable=False, primary_key=True)
+        "admin.hospitalID",  ondelete="CASCADE"), nullable=False, primary_key=True, unique=True)
     name = Column(String(128), nullable=False)
     address = Column(String, nullable=False)
     phone = Column(String, nullable=False)
     admin = relationship("Admin", back_populates="hospital")
     workers = relationship(
         "HospitalWorker", back_populates="hospital", cascade='all, delete-orphan')
+    doctors = relationship(
+        "Doctor", back_populates="hospital", cascade='all, delete-orphan')
 
 
 class HospitalWorker(BaseModel, Base):
@@ -56,6 +58,19 @@ class HospitalWorker(BaseModel, Base):
         "hospital.id",  ondelete="CASCADE"), nullable=False)
     doctors = relationship("Doctors", back_populates="worker")
     hospital = relationship("Hospital", back_populates="workers")
+
+
+class Doctor(Users):
+    """
+        doctor details
+    """
+    __tablename__ = "doctor"
+    id = Column(String, ForeignKey(
+        'user.id',  ondelete="CASCADE"), primary_key=True)
+    hospitalID = Column(String, ForeignKey(
+        "hospital.hospitalID",  ondelete="CASCADE"), nullable=False)
+    role = Column(String(50), nullable=False, default='doctor')
+    hospital = relationship("Hospital", back_populates="doctors")
 
 
 class Doctors(Users):
