--- schema/healthWorker.py
@@ -1,8 +1,0 @@
-from schema.user import User
-
-
-class HealthWorker(User):
-    hospitalID: str
-
-    class Config():
-        orm_mode = True
